feat: add showsVerticalScrollIndicator to FlatList
This is a breaking change because it changes the default value of the prop.

The reason for this change is that the default value of the prop is true,
but the default value of the prop in the native implementation is false.

This change is necessary because the default value of the prop in the native
implementation is false.